Commit v0.4.8

+++ online_store/shop/models.py
@@ -38,7 +38,7 @@ class Product(models.Model):
 class Review(models.Model):
     product = models.ForeignKey('Product', on_delete=models.CASCADE, related_name='reviews')
     fullname = models.CharField(max_length=100)
-    rating = models.FloatField()
+    rating = models.PositiveSmallIntegerField()
     text = models.TextField()
     created_at = models.DateTimeField(auto_now_add=True)
